+++ ChangeLog.md
@@ -6,6 +6,7 @@
 
 -   Added Money type in EncodeData
 -   Added close buttons in CreateEdit table dialog and entity
+-   Started work on ViewTableEntityDialog
 
 ### Changed: